+++ NEWS.md
@@ -1,3 +1,11 @@
-# fioRa (development version)
+# v0.1.1 by jan.lisec@bam.de
 
-* Initial CRAN submission.
+* Shiny App working on shinyapps.io
+* local function `predict` is working
+* spectrum plot is showing formulas and molecular structures
+* several testthat tests implemented
+* ReadMe and News.md set up
+
+# v0.0.0.9000 by jan.lisec@bam.de
+
+* initial GitHub submission